+++ src/actions.js
@@ -1,5 +1,3 @@
-import { fetchCategories, fetchRegions } from './services/api';
-
 export function getRegions(regions) {
   return {
     type: 'getRegions',
@@ -9,27 +7,11 @@ export function getRegions(regions) {
   };
 }
 
-export function loadRegions() {
-  return async (dispatch) => {
-    const regions = await fetchRegions();
-
-    dispatch(getRegions(regions));
-  };
-}
-
 export function getCategories(categories) {
   return {
     type: 'getCategories',
     payload: {
       categories,
     },
-  };
-}
-
-export function loadCategories() {
-  return async (dispatch) => {
-    const categories = await fetchCategories();
-
-    dispatch(getCategories(categories));
   };
 }
